requirements.txt
obspy>=1.4.0
numpy>=1.21.0
requests>=2.25.0
websocket-client>=1.2.0
websockets>=10.0
python-dotenv>=0.19.0